refactor(tests): split discord slash-register suite by concern
The 182 LOC test_slash_register.py held three separate test surfaces.
Move TestNormalizeName + TestBuildSignature into
test_slash_register_naming.py (89 LOC) and leave TestRegisterAll —
the public-surface behaviour — in test_slash_register.py (68 LOC).

The fake_command fixture and patch_slash_registry helper move to
discord/conftest.py so both files share them.
